refactor(nx-cmake): filterDependenciesOfProject
