+++ packages/nx-cmake/src/utils/graphUtils/filterDependenciesOfProject/detectTestFramework/detectTestFramework.ts
@@ -1,0 +1,8 @@
+import { messageIncludesCmocka } from '../messageIncludesCmocka/messageIncludesCmocka';
+import { messageIncludesGtest } from '../messageIncludesGtest/messageIncludesGtest';
+
+export const detectTestFramework = (message: string) => {
+    const includesGtest = messageIncludesGtest(message);
+    const includesCmocka = messageIncludesCmocka(message);
+    return includesGtest || includesCmocka;
+};
