+++ packages/nx-cmake/src/utils/graphUtils/filterDependenciesOfProject/getCmockaInclude/getCmockaInclude.ts
@@ -1,0 +1,7 @@
+import type { WorkspaceLayout } from '../../../../models/types';
+
+export const getCmockaInclude = (workspaceLayout: WorkspaceLayout): string => {
+    const { libsDir } = workspaceLayout;
+    const cmockaInclude = `dist/${libsDir}/cmocka/cmocka-src/include`;
+    return cmockaInclude;
+};
