+++ packages/nx-cmake/src/utils/graphUtils/filterDependenciesOfProject/getFileName/getFileName.ts
@@ -1,0 +1,10 @@
+import type { CTag } from '../../../../models/types';
+
+export const getFileName = (
+    projectRoot: string,
+    name: string,
+    tag: CTag
+): string => {
+    const fileName = `${projectRoot}/src/${name}.${tag}`;
+    return fileName;
+};
