+++ packages/nx-cmake/src/utils/graphUtils/filterDependenciesOfProject/getGtestInclude/getGtestInclude.ts
@@ -1,0 +1,7 @@
+import type { WorkspaceLayout } from '../../../../models/types';
+
+export const getGtestInclude = (workspaceLayout: WorkspaceLayout): string => {
+    const { libsDir } = workspaceLayout;
+    const gtestInclude = `dist/${libsDir}/gtest/googletest-src/googletest/include`;
+    return gtestInclude;
+};
