+++ packages/nx-cmake/src/utils/graphUtils/filterDependenciesOfProject/getWorkspaceIncludeDir/getWorkspaceIncludeDir.spec.ts
@@ -1,0 +1,8 @@
+import { getWorkspaceIncludeDir } from './getWorkspaceIncludeDir';
+
+describe('getWorkspaceIncludeDir', () => {
+    it('should return the workspace include directory', () => {
+        const result = getWorkspaceIncludeDir();
+        expect(result).toBe('include');
+    });
+});
