+++ packages/nx-cmake/src/utils/graphUtils/filterDependenciesOfProject/getWorkspaceIncludeDir/getWorkspaceIncludeDir.ts
@@ -1,0 +1,1 @@
+export const getWorkspaceIncludeDir = () => 'include';
