+++ packages/nx-cmake/src/utils/graphUtils/filterDependenciesOfProject/messageIncludesCmocka/messageIncludesCmocka.ts
@@ -1,0 +1,3 @@
+export const messageIncludesCmocka = (message: string) => {
+    return message.includes('#include <cmocka.h>');
+};
