+++ packages/nx-cmake/src/utils/graphUtils/filterDependenciesOfProject/messageIncludesGtest/messageIncludesGtest.ts
@@ -1,0 +1,3 @@
+export const messageIncludesGtest = (message: string) => {
+    return message.includes('#include <gtest/gtest.h>');
+};
