+++ packages/nx-cmake/src/utils/graphUtils/getDependenciesOfFile/getDependenciesOfFile.spec.ts
@@ -1,29 +1,6 @@
-import {
-    hasValidExtension,
-    getDependenciesOfFile,
-} from './getDependenciesOfFile';
+import { getDependenciesOfFile } from './getDependenciesOfFile';
 
-describe('hasValidExtension', () => {
-    it('should return true for valid C files with tag "c"', () => {
-        expect(hasValidExtension('file.c', 'c')).toBe(true);
-    });
-
-    it('should return true for valid C++ files with tag "cpp"', () => {
-        expect(hasValidExtension('file.cpp', 'cpp')).toBe(true);
-    });
-
-    it('should return true for valid header files with any tag', () => {
-        expect(hasValidExtension('file.h', 'c')).toBe(true);
-        expect(hasValidExtension('file.h', 'cpp')).toBe(true);
-    });
-
-    it('should return false for files with invalid extensions', () => {
-        expect(hasValidExtension('file.js', 'c')).toBe(false);
-        expect(hasValidExtension('file.txt', 'cpp')).toBe(false);
-    });
-});
-
-describe('getDependenciesOfProject', () => {
+describe('getDependenciesOfFile', () => {
     it('should return external files that start with "dist" or "include"', () => {
         const inputFiles = [
             'src/file1.h',
