+++ packages/nx-cmake/src/utils/graphUtils/getDependenciesOfFile/getDependenciesOfFile.ts
@@ -1,49 +1,7 @@
-import type {
-    ProjectFileMap,
-    ProjectGraphDependencyWithFile,
-} from '@nx/devkit';
-import type { CTag, FilteredProject } from '../../../models/types';
+import type { ProjectGraphDependencyWithFile } from '@nx/devkit';
+import type { FilteredProject } from '../../../models/types';
 import { DependencyType } from '@nx/devkit';
 import { getProjectFromFile } from '../../generatorUtils/getProjectFromFile/getProjectFromFile';
-
-export const hasValidExtension = (file: string, tag: CTag): boolean => {
-    const isCFile = file.endsWith('.c');
-    const isCppFile =
-        file.endsWith('.cpp') || file.endsWith('.cxx') || file.endsWith('.cc');
-    const isHFile = file.endsWith('.h');
-    const isHppFile =
-        isHFile ||
-        file.endsWith('.hpp') ||
-        file.endsWith('.hxx') ||
-        file.endsWith('.hh');
-    if (tag === 'c') {
-        return isHFile || isCFile;
-    } else if (tag === 'cpp') {
-        return isHppFile || isCppFile;
-    }
-    return false;
-};
-
-export const isValidProjectFile = (file: string, tag: CTag) => {
-    return (
-        !file.startsWith('dist') &&
-        !file.startsWith('include') &&
-        hasValidExtension(file, tag)
-    );
-};
-
-export const getProjectFiles = (
-    projectName: string,
-    tag: CTag,
-    fileMap: ProjectFileMap
-): string[] => {
-    const files = fileMap[projectName];
-    const projectFiles = files
-        .map(({ file }) => file)
-        .filter((file) => isValidProjectFile(file, tag));
-
-    return projectFiles;
-};
 
 export const getDependenciesOfFile = (
     mainProject: FilteredProject,
