+++ packages/nx-cmake/src/utils/graphUtils/getDependenciesOfFile/getProjectFiles/getProjectFiles.spec.ts
@@ -1,0 +1,7 @@
+import type { ProjectFileMap } from '@nx/devkit';
+import type { CTag } from '../../../../models/types';
+import { isValidProjectFile } from '../isvalidProjectFile/isValidProjectFile';
+
+describe('getProjectFiles', () => {
+    it.todo('should only get project files');
+});
