+++ packages/nx-cmake/src/utils/graphUtils/getDependenciesOfFile/isvalidProjectFile/isValidProjectFile.spec.ts
@@ -1,0 +1,6 @@
+import type { CTag } from '../../../../models/types';
+
+describe('isValidProjectFile', () => {
+    it.todo('should return true if valid project file');
+    it.todo('should return false if not valid project file');
+});
